Change argument passed to solid queue job to trigger db insert
Created a BG Job with solid queue that would create an item
in the products table of DB when triggered by Stripe webhook.

Job could not process the Stripe Object, so had to pass specific
(in this case event.id instead of the whole 'event' object)

Received 200 response and checked DB and it worked.