Refactor Stripe Webhook job to set shipping rate to Shippo Rate ID
Modify shipping options passed to Stripe Checkout to include
metadata of Shippo Rate ID (object_id).

Refactor webhook job to set Order's 'shipping_rate' to the Shippo ID
rather than the Stripe shipping rate ID created when the rate
is selected and payment processed by customer.

Stripe does NOT pass this Metadata to the Webhook event. A method in
the Webhook background job had to be created to retrieve the
Stripe Rate and then access the Shippo metadata from there.